requirements.txt
python-binance==1.0.19
pandas==2.0.3
numpy==1.24.3
PyYAML==6.0.1
schedule==1.2.0
requests==2.31.0
